+++ app/app.js
@@ -12,6 +12,7 @@ const App = Application.extend({
 	Resolver
 });
 
+console.timeEnd('ember-load');
 loadInitializers(App, config.modulePrefix);
 
 export default App;
