+++ app/routes/application.js
@@ -94,8 +94,6 @@ export default Route.extend(
 			) {
 				const adsModule = this.get('ads.module');
 
-				adsModule.init(this.get('ads.adsUrl'));
-
 				/*
 				 * This global function is being used by our AdEngine code to provide prestitial/interstitial ads
 				 * It works in similar way on Oasis: we call ads server (DFP) to check if there is targeted ad unit for a user.
