+++ package.json
@@ -45,6 +45,7 @@
   "devDependencies": {
     "bower": "1.8.2",
     "broccoli-asset-rev": "2.6.0",
+    "broccoli-babel-transpiler": "6.1.2",
     "broccoli-funnel": "2.0.1",
     "broccoli-merge-trees": "2.0.0",
     "broccoli-stew": "1.5.0",
